+++ RELEASE_NOTES.md
@@ -12,6 +12,8 @@
 * Adds `CwlEnum.symbolNames` function for getting enum symbols without any namespace prefixes
 * Adds `coerce` option to `Evaluator.evaluate` to actually perform type coercion, rather than just checking that the result is coercible to the specified type
 * Trims `StringValue` when coercing to primitive types
+* Fixes `Evaluator.finalizeInputValue` for compound and optional types
+* `Evaluator.finalizeInputValue` loads file contents from remote file source if file does not exist locally
 
 ## 0.7.4 (2021-01-05)